backend/pyproject.toml
[tool.poetry]
name = "revenda-fullstack"
version = "1.0"
description = ""
authors = ["Mateus <mateusvidotti@yahoo.com.br>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
asgiref = "^3.8.1"
django = "^5.1.2"
django-cors-headers = "^4.5.0"
djangorestframework = "^3.15.2"
djangorestframework-simplejwt = "^5.3.1"
pyjwt = "^2.9.0"
pytz = "^2024.2"
sqlparse = "^0.5.1"
psycopg2-binary = "^2.9.9"
python-dotenv = "^1.0.1"
gunicorn = "^23.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
